+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - New `utils.getPosition(state) => TablePosition` to know easily the current
   position within a table (#39)
+- New option `exitBlockType`. When provided, `Mod+Enter` will exit the current
+  table into a new block of this type.
 - Fixed cursor issues when inserting rows using Tab (#26)
 
 ## [0.10.1] - 2017-07-13